+++ projects/net.splitcells.os.state.interface/CHANGELOG.md
@@ -18,6 +18,9 @@ Changes are now written to the root changelog instead.
 
 ## Unreleased
 ### Major Changes
+* **2023-03-26** [**\#s109** `repo.process` should not create the folder of missing sub repos by default.](https://todo.sr.ht/~splitcells-net/net.splitcells.network/119)
+  If `--command-for-missing` is not set or is set as `exit 1`,
+  the folder
 * **2022-09-11** **\#168**:
   1. Rename `dependencies.test` to `test.dependencies`,
      in order to prepare `test.*` prefix for all test commands.